New algorithm for step_multidownsample().
The function default_replicates() is removed since it is not used by the new algorithm.

Rather than just repeatedly randomly downsampling the data, the new algorithm creates a maximal partition of the data's rows corresponding to the over-represented class. Each part of the partition has an equal number of rows as the under-represented class.

The resulting downsamples are less correlated and therefore give more efficient pooled predictions. In the end, fewer models need to be fit.